fix(ticket): align lifecycle hook transitions
